requirements.txt
streamlit==1.27.2
pandas==1.5.3
Pillow==9.5.0
st-aggrid==0.3.4



